[Refactor] 전체 엔티티 관계 맵핑 점검
1. 빠뜨린 부분수정
