+++ src/main/java/com/server/delivery/domain/auth/service/AuthServiceImpl.java
@@ -6,7 +6,7 @@ import com.server.delivery.common.jwt.CustomUserDetail;
 import com.server.delivery.domain.auth.dto.request.CustomerCreateRequestDto;
 import com.server.delivery.domain.auth.dto.request.OwnerCreateRequestDto;
 import com.server.delivery.domain.auth.dto.request.SignInRequestDto;
-import com.server.delivery.domain.user.dto.request.MasterSignInRequestDto;
+import com.server.delivery.domain.master.dto.request.MasterSignInRequestDto;
 import com.server.delivery.model.user.entity.User;
 import com.server.delivery.model.user.repository.UserJpaRepository;
 import com.server.delivery.util.helper.UserHelper;
